requirements.txt
pyyaml
networkx[default]